Relocate Layout component
